refactor: rebuild project architecture
