+++ src/middlewares/log-request.js
@@ -1,4 +1,4 @@
-const { logger } = require('../logger/index');
+const { logger } = require('../logger');
 
 const logRequest = (req, res, next) => {
   const { body, query, originalUrl, method } = req;
