+++ src/middlewares/start-server.js
@@ -1,0 +1,9 @@
+const startServer = (req, res, next) => {
+  if (req.originalUrl === '/') {
+    res.send('Service is running!');
+    return;
+  }
+  next();
+};
+
+module.exports = { startServer };
